requirements.txt
langchain~=0.2.16
langchain-community~=0.2.16
langchain-core~=0.2.38
langchain-openai~=0.1.23
langchain-google-genai~=1.0.10
chromadb~=0.4.24
# tiktoken is a dependency of openai and langchain, let them resolve it
pydantic>=2.7.4
python-dotenv==1.0.1
langgraph~=0.0.69
tavily-python==0.3.3
pillow==10.3.0
numpy==1.26.4
google-generativeai~=0.7.0
pytest==8.2.2
beautifulsoup4==4.12.3
lxml==5.2.2
requests>=2.32.0
openai>=1.35.10